PR feedback round 1

+++ app/src/org/commcare/android/view/c3/Configuration.java
@@ -84,7 +84,7 @@ public class Configuration {
             if (numeric.isNaN()) {
                 throw new InvalidStateException("Could not understand '" + value + "' in " + description);
             }
-            return numeric.doubleValue();
+            return numeric;
         } catch (NumberFormatException nfe) {
             throw new InvalidStateException("Could not understand '" + value + "' in " + description);
         }
